myproject/requirements.txt
Django==5.0.*
djangorestframework==3.*
django-cors-headers==4.*
tensorflow==2.* # Use the exact version you used for training if you know it
opencv-python==4.* # For image processing
Pillow # For image handling